macOS/requirements.txt
git+https://github.com/baihengaead/pywifi.git@master#egg=pywifi
PySide6>=6.7.2; sys_platform != 'darwin'
pyperclip>=1.9.0; sys_platform != 'darwin'
comtypes>=1.4.4; sys_platform != 'darwin'
pyinstaller>=6.9.0
pywin32>=306; platform_system == "Windows"
pyobjc; sys_platform == 'darwin'
pyobjc-framework-CoreWLAN; sys_platform == 'darwin'
pyobjc-framework-CoreLocation; sys_platform == 'darwin'
prettytable; sys_platform == 'darwin'
pyfiglet; sys_platform == 'darwin'